moved common ui components to folder
